Update style.css
Ended up just creating two divs in the HTML file to use aesthetically (.circle, .border). I was having a hard time using the ::before + ::after pseudo-elements and adjusting their z-indexes to stack the way I designed it so this was a fix. I also still want to change the look of the form inputs and have them not just be red + invalid immediately, but only once the user starts interacting with them and fails to match the required pattern. I'll also change the submit button to red until all inputs have been changed to valid, then have it turn green and maybe perform a small animation when clicked.