Update index.html
I added some logic to listen to every form input. If that form input in particular doesn't resolve to true using the constraint API (checkValidity()), then it gives that input a red border, and a green one if it is valid. I also added some basic regex patterns for the email/phone number, but will tweak the phone number one to be more precise.